Refactor and upgrade project structure, we have two separate services that use several pieces of shared blocks of code
