+++ app/controllers/admin/translations/index.js
@@ -1,0 +1,14 @@
+'use strict';
+
+// const TranslationPlugin = require('../../../plugins/models/translations')
+// const { APIError } = require('../../../plugins/middlewares/error');
+
+
+const feed = async function (req, res, next) {
+
+};
+
+
+module.exports = {
+  feed
+}
