+++ app/controllers/core/users/index.js
@@ -3,5 +3,5 @@
   
 module.exports = {
   ...require('./users.register.js'),
-  ...require('./users.auth.js')
+  ...require('./users.login.js')
 };
