+++ app/models/admin/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('./translations');
