+++ app/models/admin/translations.js
@@ -2,8 +2,8 @@
 
 const mongoose = require('mongoose');
 const Schema   = mongoose.Schema;
-const hash     = require('../plugins/utils/hash')
-const lastMod  = require('../plugins/models/lastMod');
+const hash     = require('../../plugins/utils/hash')
+const lastMod  = require('../../plugins/models/lastMod');
 
 
 /**
