+++ app/models/core/index.js
@@ -1,0 +1,4 @@
+'use strict';
+
+require('./oauth');
+require('./user');
