+++ app/models/core/oauth.js
@@ -2,9 +2,9 @@
 
 const mongoose = require('mongoose');
 const Schema   = mongoose.Schema;
-const config   = require('../../config');
+const config   = require('../../../config');
 
-const { tokens: tokensClient } = require('../plugins/thirdParty/redis');
+const { tokens: tokensClient } = require('../../plugins/thirdParty/redis');
 const { promisify } = require('util');
 const set    = promisify(tokensClient.set.bind(tokensClient))
 const del    = promisify(tokensClient.del.bind(tokensClient))
