--- app/models/index.js
@@ -1,6 +1,0 @@
-'use strict';
-
-require('./oauth');
-require('./user');
-require('./translations');
-require('./subtitles');
