+++ app/routes/admin/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+
+module.exports = function(app) {
+  require('./translations')(app);
+};
