+++ app/routes/admin/translations.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = function(app) {
+
+  const translations = require('../../controllers/admin/translations');
+
+
+  app.post('/api/translations', translations.feed);
+};
