+++ app/routes/public/index.js
@@ -2,8 +2,6 @@
 
 
 module.exports = function(app) {
-  require('./core')(app);
   require('./users')(app);
-  require('./translations')(app);
   require('./subtitles')(app);
 };
