+++ app/routes/public/subtitles.js
@@ -2,7 +2,7 @@
 
 module.exports = function(app) {
 
-  const subtitles = require('../controllers/subtitles');
+  const subtitles = require('../../controllers/public/subtitles');
 
 
   app.post('/api/upload', subtitles.upload);
