+++ app/routes/public/users.js
@@ -3,7 +3,7 @@
 
 module.exports = function(app) {
 
-  const users = require('../controllers/users');
+  const users = require('../../controllers/core/users');
 
 
   app.post('/users/register', users.registerUser);
