--- app/routes/translations.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = function(app) {
-
-  const translations = require('../controllers/translations');
-
-
-  app.get('/api/find', translations.findOne);
-};
